MGDOBR-168: move RhoasService to rhoas-client
