+++ manager/pom.xml
@@ -79,10 +79,6 @@
       <artifactId>actions</artifactId>
     </dependency>
     <dependency>
-      <groupId>com.redhat.service.bridge</groupId>
-      <artifactId>rhoas-client</artifactId>
-    </dependency>
-    <dependency>
       <groupId>io.quarkiverse.hibernatetypes</groupId>
       <artifactId>quarkus-hibernate-types</artifactId>
     </dependency>
