--- manager/src/main/java/com/redhat/service/bridge/manager/RhoasService.java
@@ -1,11 +1,0 @@
-package com.redhat.service.bridge.manager;
-
-import com.redhat.service.bridge.manager.models.TopicAndServiceAccount;
-
-import io.smallrye.mutiny.Uni;
-
-public interface RhoasService {
-
-    Uni<TopicAndServiceAccount> createTopicAndServiceAccount(String topicName, String serviceAccountName);
-
-}
