+++ rhoas-client/src/main/java/com/redhat/service/bridge/rhoas/RhoasService.java
@@ -1,0 +1,11 @@
+package com.redhat.service.bridge.rhoas;
+
+import com.redhat.service.bridge.rhoas.dto.TopicAndServiceAccount;
+
+import io.smallrye.mutiny.Uni;
+
+public interface RhoasService {
+
+    Uni<TopicAndServiceAccount> createTopicAndConsumerServiceAccount(String topicName, String serviceAccountName);
+
+}
